# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SleekFlow TODO API",
        "description": "Manage TODOs with recurrence schedules and task dependencies.",
        "license": {
            "name": "MIT"
        },
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register and sign in",
                "description": "`password` must be the hex-encoded SHA-256 of the user's password: the\nclient hashes it so the plaintext never crosses the wire. The server\nsalts and hashes that value again before storing it.\nReturns a bearer token, which is shown only here.",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/SessionResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Sign in",
                "description": "Takes the same hex-encoded SHA-256 as registration. Returns a bearer token.",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/SessionResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RegisterRequest": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string",
                        "maxLength": 60,
                        "minLength": 3,
                        "pattern": "^[a-zA-Z0-9._-]+$"
                    },
                    "displayName": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "password": {
                        "type": "string",
                        "minLength": 1,
                        "pattern": "^[a-f0-9]{64}$"
                    }
                },
                "required": [
                    "password",
                    "username"
                ]
            },
            "AuthenticatedUserResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "username": {
                        "type": "string"
                    },
                    "displayName": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "SessionResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "expiresAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/AuthenticatedUserResponse"
                    }
                }
            },
            "LoginRequest": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "password",
                    "username"
                ]
            },
...
        }
    }
}